requirements.txt
streamlit
python-dotenv # to store api key
langchain-community
langchain-groq